pyproject.toml
[tool.poetry]
name = "smartfarm"
version = "0.1.0"
description = ""
authors = ["blueconecell <95961799+blueconecell@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
